feat: export FeatureGateProvider and TelemetryProvider as types-only subpaths
Expose the provider contracts as dependency-free npm package subpaths so
external deployments can implement custom providers against a stable type
without importing the server's internal config schemas or zod:

  @tableau/mcp-server/features/featureGateProvider
  @tableau/mcp-server/telemetry/telemetryProvider

Each interface moves into its own leaf module (no barrel, single canonical
export site); internal consumers import the types directly from the leaf.
TelemetryAttributes is hand-written in the leaf (avoiding a zod import in the
emitted .d.ts) and kept in sync with telemetryAttributesSchema by a
bidirectional compile-time guard. A scoped tsconfig.providers.json emits the
two self-contained .d.ts, wired into all build variants. Minor version bump.
