+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@tableau/mcp-server",
   "description": "Helping agents see and understand data.",
-  "version": "2.23.0",
+  "version": "2.24.0",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/tableau/tableau-mcp.git"
@@ -20,13 +20,19 @@
   },
   "exports": {
     ".": "./build/index.js",
-    "./tracing": "./build/telemetry/tracing.js"
+    "./tracing": "./build/telemetry/tracing.js",
+    "./features/featureGateProvider": {
+      "types": "./build/features/featureGateProvider.d.ts"
+    },
+    "./telemetry/telemetryProvider": {
+      "types": "./build/telemetry/telemetryProvider.d.ts"
+    }
   },
   "scripts": {
-    "build": "tsx src/scripts/build.ts",
-    "build:desktop": "tsx src/scripts/build.ts --variant desktop",
-    "build:combined": "tsx src/scripts/build.ts --variant combined",
-    "build:dev": "tsx src/scripts/build.ts --dev",
+    "build": "tsx src/scripts/build.ts && tsc --project tsconfig.providers.json",
+    "build:desktop": "tsx src/scripts/build.ts --variant desktop && tsc --project tsconfig.providers.json",
+    "build:combined": "tsx src/scripts/build.ts --variant combined && tsc --project tsconfig.providers.json",
+    "build:dev": "tsx src/scripts/build.ts --dev && tsc --project tsconfig.providers.json",
     "build:docker": "docker build -t tableau-mcp .",
     ":build:mcpb": "npx -y @anthropic-ai/mcpb pack . tableau-mcp.mcpb",
     "build:mcpb": "run-s build:manifest :build:mcpb",
